Close + Destroy + Exit to actually terminate on restart
`Close(true)` runs the EVT_CLOSE chain (window geometry / config /
file history persistence). `Destroy()` schedules the frame for
deletion immediately rather than relying on the recursive Close
trick. `Exit()` then drops out of the main loop, which actually
terminates the process now that the debug `wxLogWindow` block is
disabled.

Also tightens `waitForPid` to `int` to match `wxProcess::Exists` /
`wxString::ToInt` and avoid the silent narrowing cast.

+++ src/lib/app/App.hpp
@@ -66,7 +66,7 @@ private:
         wxString cfgKey;               ///< `--cfg=[<category>:]<key>` (non-empty → print + exit).
         wxString loadSession;          ///< `--load-session <path>` (load + delete the file on success).
         wxArrayString files;           ///< Positional file paths.
-        long waitForPid = 0;           ///< `--wait-for-pid <pid>` (0 = no wait).
+        int waitForPid = 0;           ///< `--wait-for-pid <pid>` (0 = no wait).
         bool newWindow = false;        ///< `--new-window`.
         bool verbose = false;          ///< `--verbose`.
         bool helpRequested = false;    ///< `--help`.
